+++ assignment2/src/utils.h
@@ -20,7 +20,7 @@ enum class Status {
     ActionEnded = 6
 };
 
-double degreesToRadians(double degrees);
+float degreesToRadians(float degrees);
 std::pair<float, float> polarToCartesian(float r, float theta);
 std::vector<assignment2::Object> convertToMsgType(const std::vector<Object>& objects);
 std::vector<Object> convertToObjectType(const std::vector<assignment2::Object>& msgObjects);
